% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zine.R
\name{zine_pdf}
\alias{zine_pdf}
\alias{zine_gitbook}
\alias{zine_epub}
\title{Create zines with RMarkdown}
\usage{
zine_pdf(toc = TRUE, toc_depth = 3, highlight = "default", ...)

zine_gitbook(...)

zine_epub(...)
}
\arguments{
\item{toc}{A Boolean (TRUE or FALSE) specifying whether table of contents should be created}

\item{toc_depth}{A positive integer}

\item{highlight}{Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock". Pass NULL to prevent syntax highlighting.}
}
\value{
A zine!

A gitbook webpage

A ebook version of the zine
}
\description{
This function can be used in the YAML of Rmd files
to make zines.
}
\examples{
\dontrun{
 output: zinedown::zine_pdf
}
\dontrun{
 output: zinedown::zine_gitbook
}
\dontrun{
 output: zinedown::zine_epub
}
}
