% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinedown.R
\docType{package}
\name{zinedown}
\alias{zinedown}
\alias{zinedown-package}
\title{zinedown: A package for creating zines using R Markdown}
\description{
zinedown: A package for creating zines using R Markdown
}
\section{zine_gitbook}{

Creates an R Markdown zine template as a webpage
}

\section{zine_pdf}{

Creates an R Markdown zine template as a PDF
}

\section{zine_word}{

Creates an R Markdown zine template as a Microsoft Word document
}

\section{zine_epub}{

Creates an R Markdown zine template as an ebook
}

